requirements.txt
Pillow==8.1.0
pyscreenshot==2.3
pyTelegramBotAPI==3.7.5
